# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Trip Planner API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /trips/{tripId}/flights:
    get:
      operationId: listFlights
      tags: [flights]
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of flights
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Flight"
    post:
      operationId: createFlight
      tags: [flights]
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlightInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Flight"
  /trips/{tripId}/flights/{flightId}:
    patch:
      operationId: updateFlight
      tags: [flights]
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: flightId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlightUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Flight"
    delete:
      operationId: deleteFlight
      tags: [flights]
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: flightId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /trips/{tripId}/accommodations:
    get:
      operationId: listAccommodations
      tags: [accommodations]
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of accommodations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Accommodation"
    post:
      operationId: createAccommodation
      tags: [accommodations]
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccommodationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Accommodation"
  /trips/{tripId}/accommodations/{accommodationId}:
    patch:
      operationId: updateAccommodation
      tags: [accommodations]
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: accommodationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccommodationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Accommodation"
    delete:
      operationId: deleteAccommodation
      tags: [accommodations]
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: accommodationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /trips/{tripId}/activities:
    get:
      operationId: listActivities
      tags: [activities]
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of activities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
    post:
      operationId: createActivity
      tags: [activities]
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
  /trips/{tripId}/activities/{activityId}:
    patch:
      operationId: updateActivity
      tags: [activities]
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: activityId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
    delete:
      operationId: deleteActivity
      tags: [activities]
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: activityId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /trips/{tripId}/itinerary:
    get:
      operationId: listItineraryDays
      tags: [itinerary]
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Itinerary days
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ItineraryDay"
    post:
      operationId: createItineraryDay
      tags: [itinerary]
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItineraryDayInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItineraryDay"
  /trips/{tripId}/itinerary/{dayId}:
    patch:
      operationId: updateItineraryDay
      tags: [itinerary]
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: dayId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItineraryDayUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItineraryDay"
    delete:
      operationId: deleteItineraryDay
      tags: [itinerary]
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: dayId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthUser:
      type: object
      required: [id, username, name, role]
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, traveler]
        email:
          type: ["string", "null"]
...
    Flight:
      type: object
      required: [id, tripId, flightNumber, airline, departureAirport, arrivalAirport, departureDatetime, arrivalDatetime]
      properties:
        id:
          type: integer
        tripId:
          type: integer
        flightNumber:
          type: string
        airline:
          type: string
        departureAirport:
          type: string
        arrivalAirport:
          type: string
        departureDatetime:
          type: string
        arrivalDatetime:
          type: string
        confirmationCode:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        direction:
          type: string
          enum: [outbound, return, connecting]
    FlightInput:
      type: object
      required: [flightNumber, airline, departureAirport, arrivalAirport, departureDatetime, arrivalDatetime]
      properties:
        flightNumber:
          type: string
        airline:
          type: string
        departureAirport:
          type: string
        arrivalAirport:
          type: string
        departureDatetime:
          type: string
        arrivalDatetime:
          type: string
        confirmationCode:
          type: string
        notes:
          type: string
        direction:
          type: string
          enum: [outbound, return, connecting]
    FlightUpdate:
      type: object
      properties:
        flightNumber:
          type: string
        airline:
          type: string
        departureAirport:
          type: string
        arrivalAirport:
          type: string
        departureDatetime:
          type: string
        arrivalDatetime:
          type: string
        confirmationCode:
          type: string
        notes:
          type: string
        direction:
          type: string
          enum: [outbound, return, connecting]
    Accommodation:
      type: object
      required: [id, tripId, name, address, checkIn, checkOut]
      properties:
        id:
          type: integer
        tripId:
          type: integer
        name:
          type: string
        address:
          type: string
        checkIn:
          type: string
        checkOut:
          type: string
        type:
          type: string
          enum: [hotel, airbnb, hostel, resort, other]
        confirmationCode:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    AccommodationInput:
      type: object
      required: [name, address, checkIn, checkOut]
      properties:
        name:
          type: string
        address:
          type: string
        checkIn:
          type: string
        checkOut:
          type: string
        type:
          type: string
          enum: [hotel, airbnb, hostel, resort, other]
        confirmationCode:
          type: string
        notes:
          type: string
    AccommodationUpdate:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        checkIn:
          type: string
        checkOut:
          type: string
        type:
          type: string
          enum: [hotel, airbnb, hostel, resort, other]
        confirmationCode:
          type: string
        notes:
          type: string
    Activity:
      type: object
      required: [id, tripId, title, date]
      properties:
        id:
          type: integer
        tripId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        date:
          type: string
        time:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        type:
          type: string
          enum: [sightseeing, dining, adventure, culture, relaxation, transport, other]
        notes:
          type: ["string", "null"]
    ActivityInput:
      type: object
      required: [title, date]
      properties:
        title:
          type: string
        description:
          type: string
        date:
          type: string
        time:
          type: string
        location:
          type: string
        type:
          type: string
          enum: [sightseeing, dining, adventure, culture, relaxation, transport, other]
        notes:
          type: string
    ActivityUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        date:
          type: string
        time:
          type: string
        location:
          type: string
        type:
          type: string
          enum: [sightseeing, dining, adventure, culture, relaxation, transport, other]
        notes:
          type: string
    ItineraryDay:
      type: object
      required: [id, tripId, date, title]
      properties:
        id:
          type: integer
        tripId:
          type: integer
        date:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    ItineraryDayInput:
      type: object
      required: [date, title]
      properties:
        date:
          type: string
        title:
          type: string
        description:
          type: string
        notes:
          type: string
    ItineraryDayUpdate:
      type: object
      properties:
        date:
          type: string
        title:
          type: string
        description:
          type: string
        notes:
          type: string
...