# lib/api-zod/src/generated/types/accommodationInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Trip Planner API
 * OpenAPI spec version: 0.1.0
 */
import type { AccommodationInputType } from './accommodationInputType';

export interface AccommodationInput {
  name: string;
  address: string;
  checkIn: string;
  checkOut: string;
  type?: AccommodationInputType;
  confirmationCode?: string;
  notes?: string;
}